backend/requirements.txt
Flask>=2.0
pymongo>=3.0
pyjwt>=2.0
werkzeug>=2.0
pytest>=7.0
requests>=2.0
python-dotenv>=1.0
flask-cors>=4.0
Flask-Limiter>=2.0
